Attempted to disable the random kill of pawns in CheckForStateChange via harmony.
Looking at the logic I noticed the whole condition is based on if a random value is < num.  So I changed num to 0.

Was the easiest way for me to attack the logic via IL but now that I comprehend a few more things I think I might be able to remove the whole conditional check (faster code execution).